backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.59.9
anthropic==0.42.0
celery==5.4.0
redis==5.2.1
django-cors-headers==4.6.0
Pillow==11.1.0
tavily-python==0.3.0
exponent-server-sdk==2.2.0
jinja2==3.1.4
pytz==2024.1

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-celery-beat==2.7.0
